requirements.txt
# SQLite shim (so Chroma works on older WSL sqlite)
pysqlite3-binary==0.5.3

# App deps
streamlit==1.39.0
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==3.2.1
chromadb==0.5.11
pypdf==4.3.1
markdownify==0.12.1
beautifulsoup4==4.12.3
